првоерка

+++ tasks/chernykh_s_trapezoidal_integration/common/include/common.hpp
@@ -7,7 +7,7 @@
 
 namespace chernykh_s_trapezoidal_integration {
 
-using InType = int;
+using InType = float;
 using OutType = int;
 using TestType = std::tuple<int, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
